COMP: Remove rpoly-related code due to ACM copyright restrictions
VXL Version 6.0.0.0

The VXL major version number increased to 6 (in the root CMakeLists.txt)
because of the change to the core API (removal of vnl/algo/vnl_rpoly_roots).

The vxl core and it's dependancies should not depend on
restrictive copyrights.

Move `vnl_rpoly_roots` to new `vnl_acmalgo` module and remove
ACM-restricted code

The `vnl_rpoly_roots` implementation and related tests have been
relocated to a dedicated `vnl_acmalgo` module. All ACM
copyright-restricted code has been removed to avoid licensing issues,
ensuring compliance with VXL's practices of avoiding restrictive
copyrights.

The vxl core and it's dependancies should not depend on restrictive
copyrights.
