+++ CMakeLists.txt
@@ -65,7 +65,7 @@ foreach(p
 endforeach()
 
 project(VXL #Project name must be all caps to have properly generated VXL_VERSION_* variables
-    VERSION 5.8.0.0 # defines #MAJOR,MINOR,PATCH,TWEAK}
+    VERSION 6.0.0.0 # defines #MAJOR,MINOR,PATCH,TWEAK}
     DESCRIPTION "A multi-platform collection of C++ software libraries for Computer Vision and Image Understanding."
     LANGUAGES CXX C)
