+++ contrib/acmalgo/examples/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+link_libraries( ${VXL_LIB_PREFIX}vnl_acmalgo ${VXL_LIB_PREFIX}vnl ${VXL_LIB_PREFIX}vul )
+
+add_executable(vnl_polyroots vnl_polyroots.cxx)
