+++ contrib/acmalgo/tests/test_driver.cxx
@@ -1,0 +1,11 @@
+#include "testlib/testlib_register.h"
+
+DECLARE(test_rpoly_roots);
+
+void
+register_tests()
+{
+  REGISTER(test_rpoly_roots); # rpoly is ACM Copyright
+}
+
+DEFINE_MAIN;
