+++ contrib/acmalgo/tests/test_include.cxx
@@ -1,0 +1,7 @@
+
+
+int
+main()
+{
+  return 0;
+}
