+++ contrib/acmalgo/toms/rpoly.P
