+++ contrib/acmalgo/toms/rpoly.f
