+++ contrib/oxl/mvl/CMakeLists.txt
@@ -113,7 +113,7 @@ if(CMAKE_COMPILER_IS_GNUCXX)
 endif()
 
 vxl_add_library(LIBRARY_NAME mvl LIBRARY_SOURCES ${mvl_sources})
-target_link_libraries(mvl ${VXL_LIB_PREFIX}vil1 ${VXL_LIB_PREFIX}vnl_algo ${VXL_LIB_PREFIX}vnl ${VXL_LIB_PREFIX}vbl ${VXL_LIB_PREFIX}vgl ${VXL_LIB_PREFIX}vgl_algo ${VXL_LIB_PREFIX}vul)
+target_link_libraries(mvl ${VXL_LIB_PREFIX}vil1 ${VXL_LIB_PREFIX}vnl_algo ${VXL_LIB_PREFIX}vnl_acmalgo ${VXL_LIB_PREFIX}vnl ${VXL_LIB_PREFIX}vbl ${VXL_LIB_PREFIX}vgl ${VXL_LIB_PREFIX}vgl_algo ${VXL_LIB_PREFIX}vul)
 
 if(VXL_BUILD_EXAMPLES)
   add_subdirectory(examples)
