+++ core/vnl/algo/CMakeLists.txt
@@ -24,7 +24,6 @@ if(NETLIB_FOUND)
 # vnl_lsqr                         lsqr_
 # vnl_qr                           cqrdc_ cqrsl_ dqrdc_ dqrsl_ sqrdc_ sqrsl_ zqrdc_ zqrsl_
 # vnl_real_eigensystem             rg_
-# vnl_rpoly_roots                  rpoly_
 # vnl_sparse_symmetric_eigensystem dnlaso_
 # vnl_svd                          csvdc_ dsvdc_ ssvdc_ zsvdc_
 # vnl_svd_economy                  csvdc_ dsvdc_ ssvdc_ zsvdc_
@@ -74,7 +73,6 @@ if(NETLIB_FOUND)
     vnl_brent_minimizer.cxx vnl_brent_minimizer.h
 
     # equation solvers
-    vnl_rpoly_roots.cxx vnl_rpoly_roots.h
     vnl_cpoly_roots.cxx vnl_cpoly_roots.h
     vnl_rnpoly_solve.cxx vnl_rnpoly_solve.h
     vnl_fit_parabola.h
