+++ core/vnl/algo/tests/test_driver.cxx
@@ -22,7 +22,6 @@ DECLARE(test_qsvd);
 DECLARE(test_rational);
 DECLARE(test_real_eigensystem);
 DECLARE(test_rnpoly_roots);
-DECLARE(test_rpoly_roots);
 DECLARE(test_sparse_matrix);
 DECLARE(test_integral);
 DECLARE(test_svd);
@@ -61,7 +60,6 @@ register_tests()
   REGISTER(test_real_eigensystem);
   REGISTER(test_integral);
   REGISTER(test_rnpoly_roots);
-  REGISTER(test_rpoly_roots);
   REGISTER(test_sparse_matrix);
   REGISTER(test_svd);
   REGISTER(test_svd_fixed);
