+++ v3p/netlib/v3p_netlib_prototypes.h
@@ -272,8 +272,6 @@
 #include "minpack/qrsolv.h"
 #include "opt/lbfgs.h"
 #include "opt/lbfgsb.h"
-// #include "linalg/lsqr.h"  Removed due to ACM Copyright
-#include "toms/rpoly.h"
 #include "datapac/camsun.h"
 #include "mathews/simpson.h"
 #include "mathews/trapezod.h"
